+++ programs/BreastCancerDetector.php
@@ -3,6 +3,7 @@
 include dirname(__DIR__) . '/vendor/autoload.php';
 
 use Rubix\ML\Pipeline;
+use Rubix\ML\CommitteeMachine;
 use Rubix\ML\Datasets\Labeled;
 use Rubix\ML\Classifiers\ExtraTree;
 use Rubix\ML\NeuralNet\Layers\Dense;
@@ -13,7 +14,6 @@ use Rubix\ML\Classifiers\RandomForest;
 use Rubix\ML\NeuralNet\Optimizers\Adam;
 use Rubix\ML\Reports\MulticlassBreakdown;
 use Rubix\ML\CrossValidation\Metrics\MCC;
-use Rubix\ML\Classifiers\CommitteeMachine;
 use Rubix\ML\Classifiers\ClassificationTree;
 use Rubix\ML\Transformers\MissingDataImputer;
 use Rubix\ML\Transformers\ZScaleStandardizer;
@@ -51,8 +51,8 @@ $hidden = [
 
 $estimator = new Pipeline(new CommitteeMachine([
     [10, new MultiLayerPerceptron($hidden, 50, new Adam(0.001), 1e-4, new MCC())],
-    [8, new RandomForest(300, 0.1, 10, 3, 3, ExtraTree::class)],
-    [7, new ClassificationTree(150, 3, 5)],
+    [8, new RandomForest(300, 0.1, 10, 3, 3, 1e-4, ExtraTree::class)],
+    [7, new ClassificationTree(100, 3, 5, 1e-4)],
 ]), [
     new NumericStringConverter(),
     new MissingDataImputer('?'),
@@ -73,4 +73,4 @@ var_dump($estimator->influence());
 
 var_dump($report->generate($estimator, $testing));
 
-var_dump($estimator->proba($dataset->randomize()->head(3)));
+var_dump($estimator->predict($testing->randomize()->head(3)));