+++ src/Ensemble.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Rubix\ML;
+
+interface Ensemble extends Estimator
+{
+    /**
+     * Return the ensemble of estimators.
+     *
+     * @return array
+     */
+    public function estimators() : array;
+}