+++ tests/Classifiers/RandomForestTest.php
@@ -2,6 +2,7 @@
 
 namespace Rubix\Tests\Classifiers;
 
+use Rubix\ML\Ensemble;
 use Rubix\ML\Estimator;
 use Rubix\ML\Persistable;
 use Rubix\ML\Probabilistic;
@@ -69,6 +70,7 @@ class RandomForestTest extends TestCase
         $this->assertInstanceOf(RandomForest::class, $this->estimator);
         $this->assertInstanceOf(Classifier::class, $this->estimator);
         $this->assertInstanceOf(Multiclass::class, $this->estimator);
+        $this->assertInstanceOf(Ensemble::class, $this->estimator);
         $this->assertInstanceOf(Probabilistic::class, $this->estimator);
         $this->assertInstanceOf(Estimator::class, $this->estimator);
         $this->assertInstanceOf(Persistable::class, $this->estimator);